safari-600.7-branch/LayoutTests/fast/css3-text/css3-text-justify/getComputedStyle/getComputedStyle-text-justify-inherited-expected.txt

This test checks that the value of -webkit-text-justify is properly inherited to the child.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Value of ancestor is 'auto':
PASS window.getComputedStyle(ancestor).getPropertyCSSValue('-webkit-text-justify').cssText is 'auto'
PASS window.getComputedStyle(child).getPropertyCSSValue('-webkit-text-justify').cssText is 'auto'

Value of ancestor is 'none':
PASS window.getComputedStyle(ancestor).getPropertyCSSValue('-webkit-text-justify').cssText is 'none'
PASS window.getComputedStyle(child).getPropertyCSSValue('-webkit-text-justify').cssText is 'none'

Value of ancestor is 'inter-word':
PASS window.getComputedStyle(ancestor).getPropertyCSSValue('-webkit-text-justify').cssText is 'inter-word'
PASS window.getComputedStyle(child).getPropertyCSSValue('-webkit-text-justify').cssText is 'inter-word'

Value of ancestor is 'inter-ideograph':
PASS window.getComputedStyle(ancestor).getPropertyCSSValue('-webkit-text-justify').cssText is 'inter-ideograph'
PASS window.getComputedStyle(child).getPropertyCSSValue('-webkit-text-justify').cssText is 'inter-ideograph'

Value of ancestor is 'inter-cluster':
PASS window.getComputedStyle(ancestor).getPropertyCSSValue('-webkit-text-justify').cssText is 'inter-cluster'
PASS window.getComputedStyle(child).getPropertyCSSValue('-webkit-text-justify').cssText is 'inter-cluster'

Value of ancestor is 'distribute':
PASS window.getComputedStyle(ancestor).getPropertyCSSValue('-webkit-text-justify').cssText is 'distribute'
PASS window.getComputedStyle(child).getPropertyCSSValue('-webkit-text-justify').cssText is 'distribute'

Value of ancestor is 'kashida':
PASS window.getComputedStyle(ancestor).getPropertyCSSValue('-webkit-text-justify').cssText is 'kashida'
PASS window.getComputedStyle(child).getPropertyCSSValue('-webkit-text-justify').cssText is 'kashida'

Value of ancestor is 'inter-word, while child is 'inter-cluster':
PASS window.getComputedStyle(ancestor).getPropertyCSSValue('-webkit-text-justify').cssText is 'inter-word'
PASS window.getComputedStyle(child).getPropertyCSSValue('-webkit-text-justify').cssText is 'inter-cluster'

Value of ancestor is 'inter-ideograph, while child is 'distribute':
PASS window.getComputedStyle(ancestor).getPropertyCSSValue('-webkit-text-justify').cssText is 'inter-ideograph'
PASS window.getComputedStyle(child).getPropertyCSSValue('-webkit-text-justify').cssText is 'distribute'

PASS successfullyParsed is true

TEST COMPLETE
hello world
